Change to unit test so that it doesn't require the setup of multiple loopback devices.
Instead of using addresses in the format of `"127.0.0." + i` it now instead uses a first address of 127.0.0.1 with variable ports in the form of `50000 + i`. This way straight out of the repo you can run the tests, and to some degree achieves the same level of functional testing.

Also added a small additional bit of debug logging in `GossipService` as well as some `INFO` logging in the test - as I figure it's a test that should be suitable.